requirements.txt
pyarrow>=14.0.1
pandas>=2.0.0
numpy>=1.26.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
kafka-python>=2.0.2
requests>=2.31.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.1
matplotlib>=3.8.0
scikit-learn>=1.3.0
pytest>=7.4.0
# setuptools>=68.0.apache-airflow
# apache-airflow==2.9.1                   # in airflow container only
# apache-airflow-providers-http           # for HttpOperator
pyiceberg[pyarrow,sql-sqlite]==0.6.1     # Iceberg catalog and table management
filelock>=3.13.0                         # serialize SqlCatalog (SQLite) across parallel transforms
duckdb>=0.10                            # optional: fast local SQL over Iceberg for Gold queries
slowapi>=0.1.9
# SQLAlchemy<2.0